requirements.txt
gunicorn==20.1.0
django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2-binary
sqlparse==0.4.1
requests==2.26.0
setuptools==49.1.3
Pillow==8.3.2
mysqlclient

